style: 각 클래스 파일 분리 및 폴더 이동
