+++ Assets/Scripts/Cards/Data/CardLibraryInitializeExtensions.cs
@@ -1,6 +1,4 @@
 using Cardevil.Cards.Data.Modifiers;
-using Cardevil.Cards.Data.Modifiers.Move;
-using Cardevil.Cards.Data.Modifiers.Number;
 using Cardevil.Utils.Directions;
 using System;
 using System.Collections.Generic;
