+++ Assets/Scripts/Cards/Data/Enum/CardKind.cs
@@ -1,0 +1,4 @@
+namespace Cardevil.Cards.Data
+{
+    public enum CardKind { Attack, Move }
+}
