+++ Assets/Scripts/Cards/Data/Modifiers/ModifierType.cs
@@ -1,0 +1,12 @@
+namespace Cardevil.Cards.Data.Modifiers
+{
+    public enum ModifierType
+    {
+        AttackColor,
+        AttackDamage,
+        AttackNumSelectable,
+        AttackNumSelectableConfirm,
+        
+        MoveDirSelectable
+    }
+}
