--- Assets/Scripts/Cards/Data/Modifiers/Move.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a967b4e1ace840e2b6393551b908cd2d
-timeCreated: 1760576089
