--- Assets/Scripts/Cards/Data/Modifiers/Number.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 84bfefd9396649e880f9f2480dbea219
-timeCreated: 1760575872
