+++ Assets/Scripts/Cards/Data/Modifiers/SelectableNumberModifier.cs
@@ -1,4 +1,6 @@
-namespace Cardevil.Cards.Data.Modifiers.Number
+using Cardevil.Cards.Data.InStage;
+
+namespace Cardevil.Cards.Data.Modifiers
 {
     /// <summary>
     /// 카드에 선택 가능한 미정 값을 추가하는 Modifier.  
@@ -7,7 +9,7 @@ namespace Cardevil.Cards.Data.Modifiers.Number
     {
         public ModifierType Type => ModifierType.AttackNumSelectable;
 
-        public void Apply(BuiltCardData.Builder b)
+        public void Apply(CardData.Builder b)
         {
             b.AddNumberSelectable(null);
         }
