+++ Assets/Scripts/Cards/InStage/Model/ReadOnly/IReadOnlyStageCardsModel.cs
@@ -1,3 +1,4 @@
+using Cardevil.Cards.Data.InStage;
 using Cardevil.Cards.Data.Modifiers;
 using Cardevil.Cards.InStage.Presenter;
 using System;
@@ -35,12 +36,12 @@ namespace Cardevil.Cards.InStage.Model.ReadOnly
         /// <summary>
         /// 현재 덱의 읽기 전용 뷰.
         /// </summary>
-        IReadOnlyList<BuiltCardData> Deck { get; }
+        IReadOnlyList<CardData> Deck { get; }
         
         /// <summary>
         /// 버린 패의 읽기 전용 뷰.
         /// </summary>
-        IReadOnlyList<BuiltCardData> DiscardPile { get; }
+        IReadOnlyList<CardData> DiscardPile { get; }
 
         /// <summary>
         /// 현재 손패의 읽기 전용 뷰.
